Subject: Adjustments after second Udacity review
Body: Changes performed were as follows:

Student location changed from class to a struct. I had to create the class AnnotationObject to handle the MKAnnotation protocol requirements that were in the studentLocation. Probably there is a better way to do this.

Activity indicator is shown to be before the geocode function is called in the Info Posting View.

The unused StudentLocations array initialisation has been removed from the TabBar View Controller.

Map and pin icons have been added to the Tab Bar Controller views programatically to ensure they are displayed.

Footer: